+++ rollup.config.js
@@ -45,7 +45,7 @@ export default [{
         babel({
             extensions,
             babelHelpers: 'runtime',
-            include: ['src/**/*'],
+            include: ['src/**/*', 'node_modules/cometd/**/*'],
             presets: [
                 ['@babel/env', {
                     targets: [
@@ -57,12 +57,13 @@ export default [{
                 ['@babel/typescript'],
             ],
             plugins: [
-                '@babel/plugin-proposal-private-methods',
+                '@babel/plugin-transform-private-methods',
                 '@babel/plugin-proposal-class-properties',
                 '@babel/plugin-proposal-numeric-separator',
                 '@babel/plugin-proposal-optional-chaining',
                 '@babel/plugin-proposal-nullish-coalescing-operator',
                 '@babel/plugin-transform-runtime',
+                '@babel/plugin-transform-logical-assignment-operators',
             ],
         }),
         progress(),
@@ -123,9 +124,9 @@ export default [{
         babel({
             extensions,
             babelHelpers: 'runtime',
-            include: ['src/**/*'],
+            include: ['src/**/*', 'node_modules/cometd/**/*'],
             presets: [
-                ['@babel/env', {
+                ['@babel/preset-env', {
                     targets: [
                         'defaults',
                         'not IE 11',
@@ -135,7 +136,7 @@ export default [{
                 ['@babel/typescript'],
             ],
             plugins: [
-                '@babel/plugin-proposal-private-methods',
+                '@babel/plugin-transform-private-methods',
                 '@babel/plugin-proposal-class-properties',
                 '@babel/plugin-proposal-numeric-separator',
                 '@babel/plugin-proposal-optional-chaining',